[LOWCAR] Added ultrasonic sensor runtime_util
